Move filtermaker and deemp.h out of the top level.
filtermaker is now in scripts, and deemp.h is now part of the filter
library.
