+++ scripts/filtermaker
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 # RJS: Added amendments to work with refactored code 03-JAN-2018
