+++ tools/ld-chroma-decoder/comb.cpp
@@ -25,7 +25,7 @@
 
 #include "comb.h"
 
-#include "../../deemp.h"
+#include "deemp.h"
 
 // Public methods -----------------------------------------------------------------------------------------------------
