+++ tools/library/filter/testfilter/testfilter.pro
@@ -5,10 +5,9 @@ SOURCES += \
     testfilter.cpp
 
 HEADERS += \
+    ../deemp.h \
     ../firfilter.h \
-    ../iirfilter.h \
-    ../../../../deemp.h
+    ../iirfilter.h
 
 INCLUDEPATH += \
-    .. \
-    ../../../..
+    ..
